Debugger: Fix dumper for Qt4
Creator 14.0.0 effectively assumed "less then Qt 6 is Qt 5", even if it
is (incompatible) Qt 4.

We are now falling back to the expensive version extraction for anything
less then Qt 6  but it gets the distinction between 4 and 5 right again.

The check (once per debugger session) happens now unconditionally again
for Qt 5 and earlier and is arguably a performance regression for
Qt 5 / Creator 14 users, but it's not worse then Creator 13 where we
didn't have the shortcut mechanism

For master we should consider extending the qtVersionPing() to also
provide a cheap distinction between Qt 4 and Qt 5.

Change-Id: I4073395280fdf137e89c834bf014b9f31d74a1cd
Fixes: QTCREATORBUG-31355
Reviewed-by: Christian Stenger <christian.stenger@qt.io>