Axivion: Fix changing highlighting of issues
So far we are only fetching line markers for newly opened
documents. So, only re-opening a file updates the appearance
of text marks stored for a document.
This patch allows to remove an already set color of the mark
which in turn is responsible for highlighting the marks on
the scrollbar.
Toggling the respective setting now updates existing marks.

Change-Id: I95f65b8c48cb1f125740b88e7f744398b72ee92b
Reviewed-by: Mohammad Mehdi Salem Naraghi <mehdi.salem@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>