QmlDesigner: Remove forward declaration in model resource management
Using concepts makes it much more readable to provide actions by
parameter and not member. That removes the need of a forward declaration
of the NodeACtion class.

So now the declaration of NodeActions is automatically created by CTAD.
So adding new actions should be easier.

The NodeAction concept is providing early errors. So the interface of an
action is defined.

Cleanup forgotten metainfo.

Change-Id: Iee1bd36af556d5b753aaeb6b8ab7295f3f3378c1
Reviewed-by: Aleksei German <aleksei.german@qt.io>