QMLJSEditor: Fix calling qmlformat on save
Since the "format(...)" function would always create a temporary
file first and then let qmlformat work on that it could not find a
user supplied .qmlformat.ini in the project folder.

To fix this we now create the temporary file next to the actual file.
That way the qmlformat tool (and others) can find their config files
in the tree hierarchy.

Fixes: QTCREATORBUG-29668
Change-Id: I708efbd4bba014bc5743d901725b6fdaa06fefd8
Reviewed-by: David Schulz <david.schulz@qt.io>