QmlProjectManager: Add option for QDS compatible projects
This adds a "Create a project that you can open in QDS" checkbox. With
that being checked, the wizard uses QDS's General -> Empty application
templates to create a .qmlproject based project.

It is a replacement for the former, now defunct option for generating an
equivalent CMake based project.

QDS's templates make use of the QtQuick.Studio.Application plugin. The
former, CMake based project made that plugin available via the build
system. This new qmlproject based project relies on the plugin being
pre-installed in the selected Qt version.

Fixes: QTCREATORBUG-31355
Change-Id: I00ab0913c9ad29476070e426547b2320499195d3
Reviewed-by: Knud Dollereder <knud.dollereder@qt.io>