Qt Quick Application wizard: Remove QDS compatibility option
Generation of a Qt Quick application with the "Create a project that you
can open in QDS" option made use of QDS' original wizard templates.

Since QDS dropped generation of the CMake project via the wizard, and
the respective temple files were removed, this is not anymore possible.

This change removes the option from the wizard.

Task-number: QTCREATORBUG-31355
Change-Id: If0b9106e35c9fde708f7432106a9656a95543670
Reviewed-by: Knud Dollereder <knud.dollereder@qt.io>