Core: Adapt output window chunk size and queue timer dynamically
If we notice that our formatter takes longer to handle an incoming chunk
of output than the amount of time we wait until we handle the next one,
then we lower the chunk size and increase the interval. This way, we
ensure responsiveness in the presence of excessive output.

Task-number: QTCREATORBUG-30135
Change-Id: I1d1e31a7c6f26b50bdc048322e0f7987ddd7b317
Reviewed-by: David Schulz <david.schulz@qt.io>