QmlDesigner: Prevent deadlock in drag handling in item library
model->startDrag() starts a new event loop and doesn't exit until the
drag is over, so in the old implementation m_itemToDrag was not
cleared until the drag was complete. However, in some cases it is
possible to receive another move event while drag is still ongoing,
which would then get interpreted as start of another drag as
m_itemToDrag was still valid, causing nested drag event loops, which
caused a deadlock in gui thread.

This issue is prevented by clearing m_itemToDrag before drag starts.

Fixes: QDS-13380
Change-Id: I6395949998ede4bae08b286d3ab4a1575be82551
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>