Android: Compile app profiles before starting debugging
At https://source.android.com/docs/core/runtime/configure/art-service we
learn that:

"Starting with Android 14, on-device AOT compilation for apps (a.k.a.
dexopt) is handled by ART Service. ART Service is a part of the ART
module, and you can customize it through system properties and APIs."

This commit makes sure to have the app profiles created before the
application was started. Otherwise the service will do it in the
background and could trigger exceptions that would land in disassembly.

Fixes: QTCREATORBUG-29928
Change-Id: I5d30fa03535f03b15d5470789323c0af0246d0dd
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>