Git: InstantBlame: Simplify slot for editor setup
The slot parameter `editor` can either be an active
editor or nullptr. nullptr could either mean no
parameter was supplied or we have no current editor.

Therefore, we would have to query the current editor
again anyway, and so its easier to remove the
parameter right away and always query the current
editor.

Change-Id: Ib0495ca37863017719734bcea9bb6f61835bde95
Reviewed-by: Orgad Shaneh <orgads@gmail.com>