QmlProject: Open a fake project if a single .ui.qml file is opened
If a single ui.qml file is opened in Qt Design Studio we open
a fake project that includes the folder of the opened ui.qml file.

If a .ui.qml file is opened and no project is opened in Qt Design Studio
we open a "fake" project with the specific name, that provides a default.
We can later extend this to supporting import paths and other advanced features.

Task-number: QDS-13171
Change-Id: Icecb669d7149c4850da7dd908e9b3e1619d74c7d
Reviewed-by: Burak Hancerli <burak.hancerli@qt.io>
Reviewed-by: Thomas Hartmann <thomas.hartmann@qt.io>