Android: Fix post-build opening of package location for Qt 6.8
When building an Android package for Qt > 6.8, Qt Creator sets the env
variable QT_USE_TARGET_ANDROID_BUILD_DIR. This alters the name of the
resulting package compared to a Qt 6.7 package, which prevents the post-
build opening of package location.

This change adds a check whether the package file exists and a fallback
to its parent directory for the case it does not.

Fixes: QTCREATORBUG-22627
Change-Id: I35b63fa0a7cf19b78f460dfdb7ec35f7ec370d48
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>