Fix example categorization when only installing B2Qt
In that case the Qt examples are not registered through an actual Qt
version, but just via an additional "example set".

For these we do not have information about the actual Qt version number,
but we check if the Qt version is greater equal 6.5.1, otherwise we
default to the uncategorized view. Exclude invalid Qt version numbers
from that check to fix it.

Change-Id: I2e65856a3a3147a43c9d44c8821b85aa79b9a949
Reviewed-by: Christian Stenger <christian.stenger@qt.io>