qtcreatorcdbext: Forward PythonTargetArchDll in multi-config
In setups where no internet is available (Coin CI) we need to specify
the path to the python target arch dll.

This was done in single configuration mode via the CMake parameter
PythonTargetArchDll.

But in order to build all architectures we have now:

 * PythonTargetarm64Dll
 * PythonTargetwin32Dll
 * PythonTargetx64Dll

And Python3_ROOT_DIR is also forwarded, in order to fix one specific
host Python version.

Fixes: QTCREATORBUG-31496
Change-Id: I754720f90a4b1fc7d7434ee285c7e300dd77734e
Reviewed-by: Eike Ziller <eike.ziller@qt.io>