cdb: Match x64 target for arm64 cdb
On a Windows arm64 machine "Debugging tools for Windows" only installs
and arm64 and an x86 cdb. By default no x64 cdb is present there.

Though it is possibble to use the arm64 cdb to debug an x64 target on
these machines. Reflect that fact in our debugger matching logic.

Task-number: QTCREATORBUG-30533
Change-Id: I6f3ecb7ce393c4860e2eeac286683e2b60fea7b6
Reviewed-by: David Schulz <david.schulz@qt.io>