Revert "Debugger: simplify interrupting windows processes"
Console applications on MinGW/GDB stop when adding or removing a
breakpoint, and cannot be continued from this point.

This reverts commit 9d88da6c2bc40b17215137d9b2face5aafdfae3d.

The commit was partially reverted in 2364448f910, but looks like
this only solved some of the use-cases.

This is yet another partial revert. It recovers the old behavior
for GDB, and leaves out only CDB part.

Fixes: QTCREATORBUG-31518
Change-Id: I98c4a99fbd1e30a48731ee992e2b9ff20ac0350b
Reviewed-by: David Schulz <david.schulz@qt.io>