Change the mime database to Apache Tika
qtbase/2d2975bd9fb32bb6dc50c7f513cbe00337fc16b0
  Change the mimetype database embedded into QtCore

Like before, this is a copy from Qt, because we do not want to depend on
Qt versions for this and also always use the embedded database and never
system ones on Linux, so we get something deterministic.

Add some MIME types that are missing in the Tika database to
corresponding plugins.

Fixes: QTCREATORBUG-31272
Change-Id: I5c1ff5d7c5f22bb1276570195b0719beb5268925
Reviewed-by: David Schulz <david.schulz@qt.io>