Fix issues with Design mode and multiple editor windows
Amends 6f56310e31b8d6440e06f376fff66261b1ed0384

That change makes sure that we most of the time have a "current editor
view" that is visible. Changing the current editor view implicitly means
changing the "current editor".

When switching to a .ui file in edit mode in the main window, that
switches to Design mode. That hides the main editor view, so the code
above switches the current editor view to an external editor window, if
that exists and was ever active. Which in turn switches the current
editor which disables Design mode (if that isn't a different .ui file).
That made it impossible to open a .ui file in Design mode via switching
to it in Edit mode.

This adds a hack/workaround that disables the logic from the previous
change, if the current view is hidden and Design mode is active.

It would be better if Design mode was not so tightly coupled to the
current editor.

Fixes: QTCREATORBUG-31378
Change-Id: I9120fcda9752125eb8c7b653ca406960e7101397
Reviewed-by: David Schulz <david.schulz@qt.io>