Utils: Move the DesktopFileWatcher into own thread
Using locks makes it dangerous to wait as deadlocks may occur. Because
of that we were not able to return results from addPath and removePath.

To circumvent this we move all access into its own thread to remove that
possibility. This allows us to safely call invoke with
BlockingQueuedConnections, as the watch thread cannot be blocked from
another thread.

We add an error() function to allow more information in case
add or remove fails.

Change-Id: Idb4ff1361e8aea94f4a8f71f0e3ef74466e03879
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>