QmlDesigner: Properly parse components in BundleHelper
Instead of taking assets in the same folder as the component (which
can cause issues, for example, if the component is in the project
rather than an imported component), the component is properly parsed
(recursively) and its dependencies are collected. There are some minor
limitations due to the inherent complexity of qml. Also, some relevant
fixes and tweaks.

Fixes: QDS-13369
Fixes: QDS-13539
Fixes: QDS-13538
Change-Id: I6da9ce5cf14fb30c556fd521b6b452c3ab558f64
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>