AutoTest: Support durations for all frameworks
Even if the test framework has no builtin support for displaying
durations we are able to give at least an approximation of the
used time.
So, use the process execution duration as proximity for the
duration of all tests of an executable to avoid displaying
even more wrong summaries (when mixing frameworks with and
without duration support).
This enables the summary duration for Boost tests.

Task-number: QTCREATORBUG-31242
Change-Id: Ifa00038150cc35b66b3f20cc617e15f98b7264ca
Reviewed-by: David Schulz <david.schulz@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>