Build: Get rid of custom dependency tracking for plugins
Instead of PUBLIC_DEPENDS, our build system would keep record and track
recursive plugin dependencies manually, by attaching a custom property
to the targets that reference the targets it depends on. Then it used
DEPENDS to add all these recursive dependencies.

Get rid of this custom reimplementation of PUBLIC_DEPENDS and just use
that instead. This needs a fix that we should not separately export the
licensechecker plugin when doing a super-repository build. Otherwise
external plugins configured with the result cannot find the licensechecker
target, which is referenced by the vxworks target that is exported
together with the QtCreatorConfig.

Change-Id: I1444a0b16e5519f5310f9c0ffdfc49481ac228a4
Reviewed-by: Cristian Adam <cristian.adam@qt.io>