CMake: Fix usage of optionals in presets
Do not use optional for values that are not optional.
The `hidden` properties of configure and build presets have defaults and
are never unset. And also never checked for existence of a value. The
same for `inheritConfigureEnvironment`. In that case
the "value" in the optional was also never used, which isn't right.

Change-Id: I3dd427bd4a2434de9d2c7b5f1f16ef78dd7d1277
Reviewed-by: Cristian Adam <cristian.adam@qt.io>