TextEditor: Fix continue highlighting
The highlightStartBlock was incorrectly set to the block that was just
highlighted. This resulted in highlighting the same block over and over
again if it took longer than 20 ms to highlight. Fix this by setting the
highlightStartBlock after moving to the next block.

Additionally make sure that highlightStartBlock is set to INT_MAX after
the highlighting is done to ensure that the check of highlightStartBlock
in SyntaxHighlighterPrivate::reformatBlocks always uses the correct
start block.

Change-Id: I69ea0c7463913d55076601b6a545c127e5382685
Reviewed-by: Eike Ziller <eike.ziller@qt.io>
Reviewed-by: Christian Stenger <christian.stenger@qt.io>