ProjectExplorer: allow to avoid scanning workspace subdirectories
In order to provide file watchers for the workspace project we need a
way to completely exclude subdirectory of a workspace project. Otherwise
we might end up with an unnecessary amount of notifications for quickly
changing subdirectories like a build dir inside the workspace project.

Use the previously available "files.exclude" and fully remove the
entries of the project tree and avoid the scanning of the folders in
that list.

Change-Id: If11f62b1eabb4a7f03e4445e5f37068466feda4d
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>