Vcpkg: Make vcpkgManifests() cancelable
Hide this method in cpp.
Pass a promise, so that the function is made cancelable.
Make a promise of VcpkgManifest type (instead of returning
a list of VcpkgManifest types).
Deliver results via promise arg.
Collect results via results() method.
Remove alias for QList<VcpkgManifest>, as it's used just once now.

Change-Id: I06b76da9f40617bd6397fbee31b6db711dddcfb3
Reviewed-by: Alessandro Portale <alessandro.portale@qt.io>