QmlDesigner: Remove emitCurrentEditor hack
Sending a currentEditorChanged signal from outside of editor manager,
and without actually changing the current editor is evil. It breaks
preconditions that other parts of the code rely on.

In any case the hack no longer fixes anything, because the editor actions
no longer (411100b0378202dc617acaa236f4730eb4cc43b2) depend on the
current editor (or the change signal).

Change-Id: I819b5e483f56e2ba68bd85a9daee19b5df3da100
Reviewed-by: David Schulz <david.schulz@qt.io>