Axivion: Remove project settings
This removes the (hard) linking to a project. The user now
has to choose the preferred dashboard and the project to
inspect directly on the issues pane, similar to what is
needed for other Axivion plugins.
The last used dashboard and project will be now stored
inside the session data if its not the default session.

This breaks opening files from the issues pane if no project
or the wrong project is currently opened.
Fixing this needs further tweaking of fetching line markers
and maybe preparations for the project mapping which will be
done separately.

Change-Id: Id270c9758daa541059361407ad82adfdf6a240b5
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>