Axivion: Morph into Perspective
Replace the current approach of having separated
widgets for Axivion by a perspective tying them
together.
This gets rid of the Axivion output pane and the
mis-used navigation widget and moves their content
into the Axivion perspective.
This perspective is now part of the Debug view like
other analyzing tools.

Change-Id: I5b40c94ea32a3759fdfda41cbeaae11cc69f66cc
Reviewed-by: Mohammad Mehdi Salem Naraghi <mehdi.salem@qt.io>
Reviewed-by: hjk <hjk@qt.io>