EditorManager: Do not access SplitterOrView
How the editor area and views are organized should be considered
internal API to them. EditorManager should only need to access editor
area or view directly. This might make it possible to transform
the Design mode widget to a "different kind of editor area" in the
future,
and make it a full citizen in the editors realm.

Change-Id: I1e2ad64c9f85cee8171007e12514709487f4315e
Reviewed-by: David Schulz <david.schulz@qt.io>