Add variable expansion in tooltip
Enabled expansion of variables in CMake configuration tooltip.

Note, I honestly tried to create a unit-test for this,
but it all ended up futile: even ignoring the fact that
`ConfigModelTreeItem` uses private `InternalDataItem` type,
CMakeProjectManager does not expose some of the symbols
required for the unit test to be implemented.
At this point I do not think it is worth it to try any further.

Change-Id: I78cd52c2470225e00477ada424cb43d7da75a14c
Reviewed-by: hjk <hjk@qt.io>