Ios: Fix a possible crash on shutdown
Since IosDeviceManager is a static object, it's destructed
after the shutdown phase has passed. But since the destructor
of IosDeviceManager may be deleting the running task trees
containing the running processes, it may crash since the
ProcessReaper is gone at this stage.

To fix it, create IosDeviceManager on heap and pass a
shutdown guard as a parent, so that it's deleted before
the ProcessReaper is gone.

Fixes: QTCREATORBUG-31484
Change-Id: Ieaac16331304cffbe3f50291d7d914aebfa0b2e3
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>