QmlDesigner: Show live preview of main screen in dock view
Live preview now always shows the main screen (Screen01 by default),
regardless of which file you have open.

The live preview will display in live view instead of a floating
window if live view is enabled. It is still a separate borderless
window, it's just automatically resized and repositioned to appear
on top of live view.

Change-Id: I2c52f14dfc1ba5658eefa7ef90b441363bf7bab1
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>