Android: fix app termination logging
When an app has exited a correct log is printed that it died.
However, when the user terminates the app from within Qt Creator,
it prints both that the app was terminated and died, that is not
correct. This prints only either case when appropriate.

Also, slightly rephrase the log message in such cases, and get rid
of the starting double new lines which just show up in messed up
format with timestamp being separated from the message itself.

Change-Id: I211ff1e6d473f1dacc2cd243f64051e81a26a77a
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>