Axivion: Start using the path mapping
In case the one uses the axivion perspective without
an open project we could not open files from the issues
table without the matching project opened.
The mapping allows to find the preferred location for
a file of a project even if the project is not opened.

We explicitly give the mapping no precedence as
 * open a file without a project results in follow-up
   problems like additional syntax / semantic highlighting
   markers
 * having a project open which has the respective file
   listed is more likely to be the correct one than some
   mapping, e.g. working with different folders for
   different branches may end up opening a matching file
   from the other project folder / branch otherwise

Change-Id: I0d410f835ca2857cc2d10f09e6ba7f734dd5ff0f
Reviewed-by: Mohammad Mehdi Salem Naraghi <mehdi.salem@qt.io>
Reviewed-by: hjk <hjk@qt.io>