QmlJSEditor: Add project specific settings for qmlls
Add a new ProjectSettings class for project-specific qmlls settings.
Currently only two checkboxes exist, 'Use General Settings' to ignore the
other project specific settings, and 'Turn on' to turn qmlls on or off.

Create a QmlJSEditingProjectSettingsWidget and a
QmlJSEditingProjectPanelFactory to display the ProjectSettings.

Make sure to emit QmllsSettingsManager::instance()->settingsChanged()
when the project specific settings are changed to make sure that qmlls
project specific preferences are followed. Make
QmllsSettingsManager::useQmlls() aware of the existence of project
specific settings.

Fixes: QTCREATORBUG-31473
Change-Id: Ia8e2d03837eabe96caafba5d13728ed8e1c8615c
Reviewed-by: hjk <hjk@qt.io>