LanguageClient: fix crash in call and type hierarchy
The response to a request sent by Call or TypeHierarchy was still
handled by deleted Hierarchy instances. This happens when a second
request was sent before the first was handled. Make sure to cancel a
potentially running request when destructing a hierarchy in order to
prevent accessing the deleted instance.

Change-Id: Icd810f7a6e9eb090b28f13cb1a3254f75a27cdbb
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>