Lua: Make OptionsPage safer
If the calling script does not keep a reference to the OptionsPage
it will be garbage collected by Lua eventually. That can happen while
the SettingsDialog is open, which cannot cope with a page
being destroyed while its open.

For now we just keep the OptionsPage alive until the plugin is unloaded.

Change-Id: I8d0592731236482bbdc35078c24a5c6b962b1147
Reviewed-by: hjk <hjk@qt.io>