SquishTests: Stabilize tst_APTW03
Enabling and disabling QML debugging in a qmake-based project causes
parsing of the project. Should issues be found during parsing, those
will overwrite previously found issues.

Wait for this parsing to finish so the expected issues found during
the following build will not be overwritten by a still ongoing parse.

Change-Id: I4a41059554c10fc194674b4a529fed8ea7bdf9a6
Reviewed-by: Christian Stenger <christian.stenger@qt.io>