Utils: Introduce a dedicated Result class
Effectively wrapping a std::optional<QString> with the meaning that
"no string" is a success result and "some string" is an error result
and the string contents is an error message.

This is meant to replace uses of expected_str<void> which effectively
does the same, but has the peculiarity that a default constructed
object means "success". The default constructor of Result is private,
as "success" object Result::Ok should be used.

Change-Id: If8a01b9f97a6ecd0785f5db19980605fefa5c4b5
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>