QmlDesigner: Snap to model in 3D view
When model snapping is enabled, the dragged node snaps to the sides
of the scene axis oriented bounding volumes of models in scene space.

Since the bounding volume is always oriented to scene axes, if a cube
is rotated, it won't snap to the visible side of that cube but to the
bounding volume defined by the corners of the cube.

If dragged model has child models, those are included in the bounding
volume of the dragged node. Non-dragged models are considered
independently for snapping.
The exception is models contained in components, as typically component
is considered a single entity in the scene. Therefore each component is
has single bounding volume regardless of how many models they have
inside.

Change-Id: I5d1bb25edb57aaa0891d0097bc74e53710bdd6e7
Reviewed-by: Mahmoud Badri <mahmoud.badri@qt.io>