Do not move tooltip widget while in a layout
Attempting to move a widget when it is a part of layout
has some interesting effects.
We should only try to move the tooltip widget when it is on its own.
When the tooltip is a part of a `WidgetTip` container,
the entire container must be moved instead.
Reverted patch #576657 as no longer needed.

Fixes: QTCREATORBUG-31250
Change-Id: I448fee4deef532f35ec7159e3284464c3dcb854c
Reviewed-by: hjk <hjk@qt.io>