Android: move finished message from AndroidRunner to AndroidRunnerWorker
Not calling setFinished() when the app is terminated by the
user causes some issue and the app won't be able to restarted
again from the run button and a new deployment is needed. So
move the termination message to the worker where the "died"
message is at.

This amends 33022eec04a54c8ce87175d79738ea5c0abf7af0.

Change-Id: Ieb02c4343d0f10d988372288715fe7fc2c85a643
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>