POC: material reference (as node reference)
Implementation for POC, create children in nodes that have a property named
“materials”. The action is to create nodes that are references to materials
nodes.

The action currently behaves as if they were materials nodes, but they point
to the original nodes.

The behavior of these nodes has been changed in the Navigator view so that
they cannot be moved and their visible/alias/lock values can be modified.
They were given new actions such as editing or removing material from a node.

Change-Id: I65568ef0fd446cd179b0e8270b35c950c9967060
Reviewed-by: Miikka Heikkinen <miikka.heikkinen@qt.io>