ProjectExplorer: Consider project patterns when opening directories
In the case of opening directories as project files, we now look to see
if there is one existing project file in the directory. If so open that
project file.

This is the case for 'qtcreator .' in a directory that either contains a
'CMakeLists.txt' as a source directory, or a 'CMakeCache.txt' as a build
directory.

Fixes: QTCREATORBUG-24439
Fixes: QTCREATORBUG-30507
Change-Id: Ib7a47343369f2575782f2424b7af2e51072a9974
Reviewed-by: Christian Kandeler <christian.kandeler@qt.io>