Axivion: Allow None item for dashboard
This allows to explicitly disable the accesses of possibly
configured dashboards.

The old approach with the hard linking is gone and if the
user now accesses the Axivion perspective intentional or
not the session is tied to the dashboard until this is
changed to a different one.
That also means accessing any file inside the session may
try to request information about the file from the current
dashboard server.

Make it possible to undo the loose linking by introducing
a None item which in turn disables accessing the dashboard.

Change-Id: I04a960baad102663c4d97e652d44c86a04563416
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>