CMakePM: Build the correct executable before running
When issuing on a CMake target the context menu action "Run" the build
system would build the wrong target:
  * either "all"
  * or "Current executable" which might have been a different target

This change set makes sure that when a project target is selected in the
Project view, the build target will be restricted to this target.

This way the correct executable will be built before running.

Fixes: QTCREATORBUG-31386
Change-Id: Ibe6e50c4588ffca53ec2ac9555265023188f6e4e
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>