Axivion: Partially redo handling project selection
If we cannot find the preferred project inside the selected
dashboard we just dropped out with an error message which could
easily be missed.
Anyhow we should fetch the first available project in that case
to refill other filter controls if the dashboard contains any.
Only if the dashboard has no configured projects drop out early.

Change-Id: Iff72ccb79b9ec3e11620ae435c5322f40269c9ef
Reviewed-by: hjk <hjk@qt.io>
Reviewed-by: Mohammad Mehdi Salem Naraghi <mehdi.salem@qt.io>