ProjectExplorer: optimize file state highlight in project tree
In general file operations are way more expensive on windows compared to
linux. The current implementation of the modification state highlight of
a file in the project tree tries to get the responsible version control
system in the data function. This is called pretty often and the code
that tries to detect which version control system is used iterates over
a lot of directories. This results in notable hickups when loading and
browsing a fully expanded Qt Creator project tree here.

Reduce the number of file operations by caching the file state inside
the FileNode.

Change-Id: I7d7a82b296494f964ae9259c2e10c1da2ae4aa3a
Reviewed-by: Orgad Shaneh <orgads@gmail.com>