ExtensionManager: Change position of the short description
The short description is currently shown as initial part of the
description, followed by the long description.

With this change, the long description is shown by its own. And the
short description moves to other places: on the bottom of the card
(replacing the tags) and on the bottom of the details-top element.

Fixes: QTCREATORBUG-31813
Change-Id: I927573b9699ad830896fdebec4077aadf2df7b9e
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>