Android: Require manual close after installation is complete
When there were some issues during installation, marked with
red lines, it could happen that the user haven't got a chance to
read them, as the dialog was automatically closed after
the installation was done.

To fix it, hide Cancel button on installation finish, show
Close button instead and prolong the visibility of the dialog until
the Close button is pressed.

Task-number: QTCREATORBUG-31812
Change-Id: I63287150b293d3ea1a696a7925bfaa1983e00ae0
Reviewed-by: Assam Boudjelthia <assam.boudjelthia@qt.io>