Theme: Implement QIcon::On state for new themes
The spec for the new themes defines that pressed toolbar and mode
buttons have a stronger color. Defined is the change from
Token_Text_Muted to Token_Text_Default.

Themes which define another color than Token_Text_Muted for icons (it is
the case for all other existing themes), are not affected by all of
this.

This change leads Utils::Icon to create an QIcon::On version of icons
using the above mentioned color change. The actual painting of the new
QIcon::On state is also implemeted for FancyToolBar and in
ManhattanStyle.

Fixes: QTCREATORBUG-31826
Change-Id: Ib57f1cbc14839cba6303220128f589883c990b9e
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>