iOS: Do not post message on fallback transfer as error
Amends 84cc7208dee3a8176ccbf24f7f46167c680eed7a

Since that change, "error" messages from iostool are posted as errors in
the Issues view. The problem with that is that iostool did not
distinguish between errors and informative messages, so the message
"Failed to transfer and install application, trying old way"
was added as an error. Errors in the Issues view also make Qt Creator
pop up the question if the user wants to continue with starting the
application even though there are errors, which is wrong and annoying
for that message.

Split messages from iostool into errors and "normal" messages. Normal
messages are not posted in the Issues view and use the "normal" font for
the text in the output view.

Fixes: QTCREATORBUG-31811
Change-Id: I99b456b408585440028f5211195d205cfeb54f09
Reviewed-by: Marcus Tillmanns <marcus.tillmanns@qt.io>