Revert "Utils: Expand FilePaths per part, not as a full string"
This reverts commit e8855f97d59670f8be8103bbe662de761a225f39.

Reason for revert: Since no one remembers what potential side effects
there may have been, and the commit in general makes "expand" worse.

The problem with e8855f97d59670f8be8103bbe662de761a225f39 is when
"fileNameWithVariables" contains something like "%{ProjectDirectory}",
which expands to e.g. `docker://image/some/path`, we cannot rely on
"fromParts" to do the right thing. It will write everything into the
corresponding slot of a FilePath no matter what the content is, so the
"docker://" wil become part of the path, and not be recognized as the
scheme.

Change-Id: I72aeba0ed6c4a6c9a9e5ecb52b3e8e4043ae5669
Reviewed-by: hjk <hjk@qt.io>